Explicitly install node-red@0.20.0 inside the CI/CD builds
Although this step is clearly listed in the docs, this caught me out
when running through the examples because, at the time of writing,
non-0.18.0/0.20.0 are the only versions that appear to work with this
framework, not latest. Whilst it is clearly expressed as an npm
peerDependency and using the most recent version of node-red does print
errors about this, new people to the framework may be put off by this -
as it *appear* that it doesn't work out-of-the-box.

Part of the reason for this PR is to demonstrate how to use these tools.